requirements.txt
click==8.1.7
colorama==0.4.6
colorthief==0.2.1
darkdetect==0.8.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
PyQt-Fluent-Widgets==1.5.4
PyQt5==5.15.9
PyQt5-Frameless-Window==0.3.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
python-dotenv==1.0.1
pywin32==306
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
scipy==1.12.0
ultralytics==8.1.43
## hth
aiofiles==23.2.1
altair==5.0.1
altgraph==0.17.3
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
attrs==23.1.0
auto-py-to-exe==2.37.0
blind-watermark==0.4.4
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
decorator==4.4.2
easygui==0.98.3
Eel==0.16.0
exceptiongroup==1.1.3
fastapi==0.101.1
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.42.1
fsspec==2023.6.0
future==0.18.3
gevent==23.7.0
gevent-websocket==0.10.1
gradio_client==0.5.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-resources==6.0.1
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.7.2
moviepy==1.0.3
MyQR==2.3.1
Nuitka==1.7.10
ordered-set==4.1.0
orjson==3.9.5
packaging==23.1
pandas==2.0.3
pefile==2023.2.7
platformdirs==3.10.0
proglog==0.1.10
psutil==5.9.5
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.7
pyparsing==3.0.9
pypng==0.20220715.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytools==2023.1.1
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzbar==0.1.9
qrcode==7.4.2
qt-material==2.14
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
toolz==0.12.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
websockets==11.0.3
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.0
zstandard==0.21.0
zxing==1.0